Fixes to get comparisons to work.

+++ repeng/probes/contrast_consistent_search.py
@@ -34,7 +34,7 @@ class CcsProbe(torch.nn.Module, BaseProbe):
         self,
         activations: Float[np.ndarray, "n d"],  # noqa: F722
     ) -> PredictResult:
-        probabilities = self(torch.tensor(activations)).numpy()
+        probabilities = self(torch.tensor(activations, dtype=torch.float32)).numpy()
         return PredictResult(
             logits=probabilities,
             labels=probabilities > 0.5,
